Adding (some) extensibility to define custom codecs for client and server
