+++ eureka-core/src/main/java/com/netflix/eureka/AbstractInstanceRegistry.java
@@ -102,6 +102,7 @@ public abstract class AbstractInstanceRegistry implements InstanceRegistry {
 
     private static final AtomicReference<EvictionTask> EVICTION_TASK = new AtomicReference<EvictionTask>();
 
+    private final ResponseCache responseCache;
 
     /**
      * Create a new, empty instance registry.
@@ -112,6 +113,7 @@ public abstract class AbstractInstanceRegistry implements InstanceRegistry {
         deltaRetentionTimer.schedule(getDeltaRetentionTask(),
                 EUREKA_CONFIG.getDeltaRetentionTimerIntervalInMs(),
                 EUREKA_CONFIG.getDeltaRetentionTimerIntervalInMs());
+        responseCache = new ResponseCache();
     }
 
     public long getLocalRegistrySize() {
@@ -120,6 +122,10 @@ public abstract class AbstractInstanceRegistry implements InstanceRegistry {
             total += entry.size();
         }
         return total;
+    }
+
+    public ResponseCache getResponseCache() {
+        return this.responseCache;
     }
 
     /**
@@ -1159,7 +1165,7 @@ public abstract class AbstractInstanceRegistry implements InstanceRegistry {
 
     private void invalidateCache(String appName, @Nullable String vipAddress, @Nullable String secureVipAddress) {
         // invalidate cache
-        ResponseCache.getInstance().invalidate(appName, vipAddress, secureVipAddress);
+        responseCache.invalidate(appName, vipAddress, secureVipAddress);
     }
 
     private static final class RecentlyChangedItem {
